pyproject.toml
[project]
name = "docx-replacement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-docx>=1.1.0"]
